version3 website (PH Themed)/requirements.txt
amqp==5.2.0
anyio==4.6.2.post1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.1
django-crontab==0.7.1
djangorestframework==3.15.2
exceptiongroup==1.2.2
googlemaps==4.10.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
kombu==5.4.2
pillow==10.4.0
prompt_toolkit==3.0.48
Pygments==2.18.0
pyTelegramBotAPI==4.23.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.6
requests==2.32.3
schedule==1.2.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
